add sanity check to ensure GitHub Updater is active

+++ clear-github-updater-cache.php
@@ -4,7 +4,7 @@
  * Plugin URI:        https://github.com/afragen/clear-github-updater-cache
  * Description:       Deletes the GitHub Updater options from the database.
  * Author:            Andy Fragen
- * Version:           0.1
+ * Version:           0.2
  * Author URI:        https://github.com/afragen/
  * GitHub Plugin URI: https://github.com/afragen/clear-github-updater-cache
  * GitHub Branch:     force-check
@@ -12,6 +12,8 @@
 
 
 add_action( 'plugins_loaded', function() {
-	include_once __DIR__ . '/src/Purge_Cache.php';
-	new Purge_Cache();
+	if ( class_exists( 'Fragen\\GitHub_Updater\\Base' ) ) {
+		include_once __DIR__ . '/src/Purge_Cache.php';
+		new Purge_Cache();
+	}
 } );
